% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn_only.R
\name{burn_only}
\alias{burn_only}
\title{Burn-in only}
\usage{
burn_only(
  priors,
  surv = 0.5,
  rep_m,
  rep_sd,
  males_only = TRUE,
  burn = 100,
  n_iter = 100,
  n_cores = 4
)
}
\arguments{
\item{priors}{A dataframe of priors with the follow columns: ini_pop, ini_syls, innov, dem, p_att, a, v. Each row corresponds to an iteration of the model.}

\item{surv}{Average overwinter survival.}

\item{rep_m}{Mean repertoire size across observed dataset.}

\item{rep_sd}{Standard deviation of repertoire size across observed dataset.}

\item{males_only}{Whether only males are included in simulation (TRUE/FALSE).}

\item{burn}{Length of burn-in phase.}

\item{n_iter}{Number of iterations.}

\item{n_cores}{Number of cores for parallelization.}
}
\value{
Returns the new repertoire of the learner.
}
\description{
The burn-in component of the agent-based model of cultural transmission.
}
