% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{learn}
\alias{learn}
\title{Learn}
\usage{
learn(rep_size, reps, locs, loc, t_x, syl_counter, num_dems, a, innov, p_att)
}
\arguments{
\item{rep_size}{Learner repertoire size.}

\item{reps}{Demonstrator repertoires.}

\item{locs}{Demonstrator locations.}

\item{loc}{Learner location.}

\item{t_x}{Demonstrator attractiveness.}

\item{syl_counter}{Syllable counter.}

\item{num_dems}{Number of demonstrators.}

\item{a}{Level of conformity bias.}

\item{innov}{Innovation rate.}

\item{p_att}{Proportion of attractive syllables.}
}
\value{
Returns a one matrix: the Simpson's diversity from each burn-in year (column) and iteration (row).
}
\description{
An internal function that simulates learning using C++.
}
